requirements.txt
# conda install pytorch torchvision torchaudio pytorch-cuda=12.4 -c pytorch -c nvidia

jupyter
jupyterlab
numpy
pandas
pillow
matplotlib
seaborn
tensorboard
lightning
scikit-learn
scikit-image
opencv-python
loguru
hydra-core
